requirements.txt
django==5.2.4
djangorestframework==3.16.0
pandas==2.3.0
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
django-cors-headers==4.3.1
dj-database-url==2.1.0
psycopg2-binary==2.9.7
